#1297 Move fixtures to conftest.py, declare fixtures using @pytest.mark.usefixtures and avoid passing patched objects as arguments if they won't be used
